Refactor code: divide CompanyInfoContent into smaller components
